Add files via upload
To identify potential misrepresentation, I developed a two competing predictive machine-learning models based on other variables available on the respondents, namely Xgboost and Random Forest Models. This approach is based on the intuition that honest responses will be consistent with the other information we have on the respondents. This means, therefore, that significant differences between predicted statuses and actual survey responses may be indicative of a risk of misrepresention, which could then be used to inform other schemes for validating the accuracy of their responses ( for example, medical tests or medical records).